V1

+++ Core/Model.php
@@ -2,32 +2,34 @@
     class Model
     {
         public $bdd = null;
-        protected $controller ;
-        function __construct($ctr=null) {
+        //protected $controller ;
+        function __construct() {
             $tmp = Conf::$conf[Conf::$name];
             $this->bdd = new PDO("mysql:host=".$tmp["host"].";dbname=".$tmp["dbname"]."",$tmp["username"], $tmp["password"],
             array(PDO::MYSQL_ATTR_INIT_COMMAND => 'SET NAMES utf8'));
             $this->bdd->setAttribute(PDO::ATTR_ERRMODE, PDO::ERRMODE_EXCEPTION);
-            if($ctr != null && isset($ctr)) $this->controller = $ctr ;
+            //if($ctr != null && isset($ctr)) $this->controller = $ctr ;
         }
-        public static function getBDD(){
-            $mod = new Model();
-            return $mod->bdd;
-        }
-        public static function Get($query){
-            $req = Model::getBDD()->prepare($query);
+        public function Get($query){
+            //Func::debug($query);
+            $req = $this->bdd->prepare($query);
             $req->execute();
             return $req->fetchAll(PDO::FETCH_ASSOC);
-        } 
-        public static function Exec($query){
-            $db = Model::getBDD();
-            $stmt = $db->prepare($query);
+        }
+        public function Exec($query){
+            $stmt = $this->bdd->prepare($query);
             $stmt->execute();
-            if(strpos(strtolower($query),"insert") !== false) return $db->lastInsertId();
+            if(strpos(strtolower($query),"insert") !== false) return $this->bdd->lastInsertId();
             return 0;
         }
         public function Delete($id){
-            $this->Exec("DELETE FROM ".get_class($this)." WHERE id = ".$id);
+            if(AuthUser::IsAdministrator()) $this->Exec("DELETE FROM ".get_class($this)." WHERE id = ".$id);
+        }
+        public function FindAll(){
+            return $this->Get("SELECT * FROM ".get_class($this));
+        }
+        public function FindOne($id){
+            return $this->Get("SELECT * FROM ".get_class($this)." WHERE id = '".$id."'")[0];
         }
     }
 ?>
